requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-community
python-dotenv
requests
beautifulsoup4
groq
